Moved toggle-follow to a new user router
