+++ src/app.ts
@@ -5,6 +5,7 @@ import cors from 'cors';
 import { authRouter, chatRouter, commentRouter, groupRouter, postRouter } from './routes';
 import { authMiddleware, errorMiddleware, loggerMiddleware } from './middlewares';
 import { getCorsOptions } from './utils';
+import { userRouter } from './routes/user.router';
 
 dotenv.config();
 
@@ -16,6 +17,7 @@ app.use(bodyParser.json());
 
 app.use(loggerMiddleware);
 app.use('/auth', authRouter);
+app.use('/users', authMiddleware, userRouter);
 app.use('/chats', chatRouter);
 app.use('/groups', authMiddleware, groupRouter);
 app.use('/posts', authMiddleware, postRouter);
