+++ src/controllers/index.ts
@@ -3,3 +3,4 @@ export * as ChatController from './chat.controller';
 export * as CommentController from './comment.controller';
 export * as GroupController from './group.controller';
 export * as PostController from './post.controller';
+export * as UserController from './user.controller';
