+++ src/routes/user.router.ts
@@ -1,0 +1,12 @@
+import { Router } from 'express';
+import { validationMiddleware } from '../middlewares';
+import { userIdSchema } from '../validations';
+import { UserController } from '../controllers';
+
+export const userRouter: Router = Router();
+
+userRouter.post(
+  '/toggle-follow/:userId',
+  validationMiddleware({ paramsSchema: userIdSchema }),
+  UserController.toggleFollow,
+);
